move config related peripherals to config module, fix tests
